README.md: fork scope and goals + Nespresso motto (yep!)
Hard times when we are forced to explain the obvious, again & again.

On my own way translated in Italian is "a modo mio" and yep!, it is
the Nespresso (a well-known Italian brand for a pod coffee machine)
motto plus the name of a family of their products. Not just beer! ;-)

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>