gEeprom initialisation as extern requires type casting (!!)
The user github/rob-w with his pull request #200:

-> https://github.com/armel/uv-k5-firmware-custom/pull/200

Shown that avoiding structures to deal with the EEPROM saves important
space into the firmware image. In particular that space is as large as
the structure size. In the case of @rob-w it was 16 bytes.

In case of EEPROM_Config_t the size is much much bigger than 16 bytes.

As wrote by e-mail to Terri Kennedy the woman behyond, the Baofeng BF-H8
Pro GPS firmware, on 21.oct.2024, we are seen SOMETHING here!

<<Not in particular because of the code, but because eeprom and the
related state machine.>> and the waste of pre-allocated data in the FW.

By the way, the aim of this commit about to solve this warning notice:

make clean; make -s ENABLE_SPECTRUM=1 ENABLE_FMRADIO=1

/app/settings.h:287:24: warning: type of 'gEeprom' does not match original declaration [-Wlto-type-mismatch]
  287 | extern EEPROM_Config_t gEeprom;
      |                        ^
settings.c:39:17: note: 'gEeprom' was previously declared here
   39 | EEPROM_Config_t gEeprom = { 0 };
      |                 ^
settings.c:39:17: note: code may be misoptimized unless '-fno-strict-aliasing' is used
   text    data     bss     dec     hex filename
  60548      52    3388   63988    f9f4 f4hwn

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>